Move legend markers to a tooltip and heading inside the plot to make better use of space
