+++ src/qlever/monitor_queries/widgets/detail_switcher.py
@@ -1,4 +1,4 @@
-"""Bottom pane that holds the SPARQL view and the resource plot with heading.
+"""Bottom pane that holds the SPARQL view and the resource plot.
 
 Both panes stay mounted and one is shown, so switching keeps the other's
 state: the plot its rolling timer, the SPARQL view its scroll position.
@@ -12,7 +12,7 @@ from textual.app import ComposeResult
 from textual.widgets import ContentSwitcher
 
 from qlever.monitor_queries.models import ResourcePlot, SparqlContent
-from qlever.monitor_queries.widgets.plot_with_heading import PlotWithHeading
+from qlever.monitor_queries.widgets.resource_plot_pane import ResourcePlotPane
 from qlever.monitor_queries.widgets.sparql_pane import SparqlPane
 
 SPARQL_ID = "sparql-pane"
@@ -42,7 +42,7 @@ class DetailSwitcher(ContentSwitcher):
 
     def compose(self) -> ComposeResult:
         yield SparqlPane(id=SPARQL_ID)
-        yield PlotWithHeading(
+        yield ResourcePlotPane(
             self.source, self.refresh_interval, self.reload, id=PLOT_ID
         )
 
@@ -60,4 +60,4 @@ class DetailSwitcher(ContentSwitcher):
 
     def replot(self) -> None:
         """Redraw the plot; a no-op while the SPARQL pane is shown."""
-        self.query_one(PlotWithHeading).replot()
+        self.query_one(ResourcePlotPane).replot()
